refactor: 将路由功能重构为蓝图结构
将原app.py中的路由功能拆分到system_blueprint下的多个模块
修复hitokoto.html中表单按钮位置错误
更新auth_blueprint中的重定向路径以匹配蓝图结构
