+++ templates/hitokoto.html
@@ -97,11 +97,10 @@
             </div>
         </div>
         <input name="csrf_token" type="hidden" value="{{ csrf_token() }}">
+        <button type="submit" class="btn btn-primary w-100">💾提交</button>
 </form>
 </div>
 </div>
-        <button type="submit" class="btn btn-primary w-100">💾提交</button>
-    </form>
 </body>
 
 </html>
